requirements.txt
#pip install -r requirements.txt  --para instalar lo de este archivo

django
pillow
django-admin-interface
django-crispy-forms
mysql
mysqlClient
requests

##Cosas que intalar para mysql
wheel
#mysqlclient-1.4.6-cp38-cp38-win32.whl

##Cosas que intalar para el api
djangorestframework
markdown
django-filter  
     
#npm install -g json-server